requirements.txt
dvc==2.8.3
joblib==1.2.0
jupyter==1.0.0
jupyter_contrib_nbextensions==0.7.0
matplotlib==3.7.0
numpy==1.24.2
pandas==1.5.3
pytest==7.2.1
python-box==7.0.1
pyyaml==6.0
scikit-learn==1.2.1
scipy==1.10.1
tqdm==4.64.1
mlxtend==0.21.0
pyECLAT==1.0.2
plotly==5.13.1
PyYAML==6.0
openpyxl==3.1.2